Increase to .NET 7.0 & Add ugly publishing scripts
* Scripts got added for DamageCalculator and ConfigManager
* Batch scripts will need 7z be added to PATH, as well as the two specified profiles in order to work
* Alter publish profiles to fit to the scripts, one for self-contained and one for lower file size
* Add explicit cast that is now needed to remove ambiguity

+++ SteamShared/SteamShared/SteamShared/Models/NavArea.cs
@@ -21,7 +21,7 @@ namespace SteamShared.Models
             this.ID = reader.ReadUInt32();
 
             if (navVersion <= 8)
-                this.AttributeBitField = BitConverter.GetBytes(reader.ReadByte());
+                this.AttributeBitField = BitConverter.GetBytes((short)reader.ReadByte()); // .NET 6 would cast the byte to a short, .NET 7 will be ambiguous between short and Half
             else if (navVersion <= 12)
                 this.AttributeBitField = BitConverter.GetBytes(reader.ReadUInt16());
             else if (navVersion >= 13)
